requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
chardet==3.0.4
discord.py==1.3.4
idna==2.10
multidict==4.7.6
pendulum==2.1.1
python-dateutil==2.8.1
pytzdata==2020.1
six==1.15.0
websockets==8.1
yarl==1.4.2